# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/attendee/register": {
      "post": {
        "operationId": "AttendeeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttendeeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attendee created successfully!"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "summary": "Register an attendee",
        "tags": [
          "Attendee"
        ]
      }
    },
...
    "/api/v1/attendee/{id}": {
      "get": {
        "operationId": "AttendeeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Attendee"
        ]
      },
      "patch": {
        "operationId": "AttendeeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAttendeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Attendee"
        ]
      },
      "delete": {
        "operationId": "AttendeeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Attendee"
        ]
      }
    },
    "/api/v1/users/create": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "summary": "Create a new user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/login": {
      "post": {
        "operationId": "UsersController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully"
          },
          "401": {
            "description": "Invalid Credentials"
          }
        },
        "summary": "Login user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/verifyEmail/{token}": {
      "get": {
        "operationId": "UsersController_verifyEmail",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Users retrieved successfully"
          },
          "404": {
            "description": "No records found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/registrations/register": {
      "post": {
        "operationId": "RegistrationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Registrations"
        ]
      }
    },
...
    "/api/v1/registrations/{id}": {
      "get": {
        "operationId": "RegistrationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Registrations"
        ]
      },
      "patch": {
        "operationId": "RegistrationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Registrations"
        ]
      },
      "delete": {
        "operationId": "RegistrationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Registrations"
        ]
      }
    },
    "/api/v1/attendee-group/register": {
      "post": {
        "operationId": "AttendeeGroupController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAttendeeGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered group attendance successfully"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "summary": "Register group attendee",
        "tags": [
          "Attendees Group"
        ]
      }
    },
...
    "/api/v1/attendee-group/{id}": {
      "get": {
        "operationId": "AttendeeGroupController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Attendees Group"
        ]
      },
      "patch": {
        "operationId": "AttendeeGroupController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAttendeeGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Attendees Group"
        ]
      },
      "delete": {
        "operationId": "AttendeeGroupController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Attendees Group"
        ]
      }
    }
  },
  "info": {
    "title": "PlusUrEvents Backend",
    "description": "API Documentation for PlusUrEvents application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://plusureventsbackend.vercel.app/",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
      "CreateAttendeeDto": {
        "type": "object",
        "properties": {
          "lastname": {
            "type": "string",
            "example": "Johnson"
          },
          "firstname": {
            "type": "string",
            "example": "Kelvin"
          },
          "email": {
            "type": "string",
            "example": "example@gmail.com"
          },
          "phone": {
            "type": "string",
            "example": "2022222222"
          },
          "residence": {
            "type": "string",
            "example": "Legon, Accra"
          },
          "firstTime": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "lastname",
          "firstname",
          "email",
          "phone",
          "residence",
          "firstTime"
        ]
      },
      "UpdateAttendeeDto": {
        "type": "object",
        "properties": {}
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John"
          },
          "email": {
            "type": "string",
            "example": "john@mail.com"
          },
          "username": {
            "type": "string",
            "example": "johny"
          },
          "password": {
            "type": "string",
            "example": "Kei234#a%o"
          }
        },
        "required": [
          "name",
          "email",
          "username",
          "password"
        ]
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateRegistrationDto": {
        "type": "object",
        "properties": {
          "event_id": {
            "type": "number",
            "example": 2
          },
          "responses": {
            "type": "object",
            "example": {
              "1": "John Doe",
              "2": "john@email.com",
              "3": "M"
            },
            "description": "Key is field_id, value is the response"
          }
        },
        "required": [
          "event_id",
          "responses"
        ]
      },
      "UpdateRegistrationDto": {
        "type": "object",
        "properties": {
          "event_id": {
            "type": "number",
            "example": 2
          },
          "responses": {
            "type": "object",
            "example": {
              "1": "John Doe",
              "2": "john@email.com",
              "3": "M"
            },
            "description": "Key is field_id, value is the response"
          }
        }
      },
      "CreateAttendeeGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": ""
          },
          "address": {
            "type": "string",
            "example": ""
          },
          "contact_person_name": {
            "type": "string",
            "example": ""
          },
          "contact_person_phone": {
            "type": "string",
            "example": ""
          },
          "contact_person_email": {
            "type": "string",
            "example": ""
          },
          "number_heads": {
            "type": "number",
            "example": ""
          }
        },
        "required": [
          "name",
          "address",
          "contact_person_name",
          "contact_person_phone",
          "contact_person_email",
          "number_heads"
        ]
      },
      "UpdateAttendeeGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": ""
          },
          "address": {
            "type": "string",
            "example": ""
          },
          "contact_person_name": {
            "type": "string",
            "example": ""
          },
          "contact_person_phone": {
            "type": "string",
            "example": ""
          },
          "contact_person_email": {
            "type": "string",
            "example": ""
          },
          "number_heads": {
            "type": "number",
            "example": ""
          }
        }
      }
    }
  }
}